+++ src/spinner.rs
@@ -3,8 +3,6 @@ use tui::{backend::Backend, buffer::Cell, Terminal};
 
 static SPINNER_CHARS: &[char] = &['|', '/', '-', '\\'];
 
-///
-#[derive(Default)]
 pub struct Spinner {
     idx: usize,
 }
@@ -14,6 +12,10 @@ impl Spinner {
     pub fn update(&mut self) {
         self.idx += 1;
         self.idx %= SPINNER_CHARS.len();
+    }
+
+    pub fn new() -> Self {
+        Self { idx: 0 }
     }
 
     /// draws or removes spinner char depending on `pending` state