+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- support for color themes and light mode([#28](https://github.com/extrawurst/gitui/issues/28))
 
 ## [0.2.6] - 2020-05-18
 ### Fixed